# Conflicts:
#	src/main/java/ru/skypro/homework/controller/AdsController.java

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -12,5 +12,5 @@ import java.util.List;
  */
 @Repository
 public interface AdRepository extends JpaRepository<AdEntity,Long> {
-    List<AdEntity> findAllByAuthorId(int id);
+    List<AdEntity> findAllByAuthorId(Long id);
 }
